// frontend/src/app/interfaces/common.interface.ts
export interface User {
  id: number;
  name: string;
  company: string;
  site: string;
  telephone: string;
  email: string;
  balance: string;
  balancePartner: string;
  registerIp: string;
  lastLoginIp: string;
  addedDateTime: string;
  comment: string;
  isServicePaused: number;
  invitedByUserId: number;
  isPartner: number;
  enablePartnerPayments: number;
  wmr: string;
  timezone: string;
  language: string;
  languages: string[];
  accessToken?: string;
  isLoggedIn: boolean;
  isAdmin: boolean;
  currency: string;
  config: any;
  extensionSettings: UserExtensionSettings;
  notifications: UserNotification[];
  paidUntilDateTime: string;
  partnerPercent: string;
  isPaid: boolean;
  tariff: Tariff;
  tariffAmount: number;
  tariffPeriod: string;
  penaltyAmount: number;
  languageLevel: string;
  favoriteCategoryId: number[];
  dailyGoal: number;
  avatar: string;
  password?: string;
}

export enum Tariff {
  FREE = 'free',
  MONTH = 'monthly',
  YEAR = 'yearly'
}

export interface UserNotification {
  id: string;
  color: string;
  title: string;
  text: string;
}